+++ examples/wsgi_api.py
@@ -1,0 +1,4 @@
+from sams.apps.api.app import get_app
+from api import settings
+
+application = get_app(__name__, config=settings)