+++ examples/wsgi_file_server.py
@@ -1,0 +1,4 @@
+from sams.apps.file_server.app import get_app
+from api import settings
+
+application = get_app(__name__, config=settings)