+++ src/clients/python/sams_client/__init__.py
@@ -11,5 +11,5 @@
 
 from .client import SamsClient # noqa
 
-__version__ = '0.2.2'
+__version__ = '0.2.3'
 __all__ = 'SamsClient'